+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - The s3sync sidecar container to be Fargate 1.4.0 compatible
 - Added ability to export datacut queries to CSV from admin listing
+- The JupyterLab Python container to be Fargate 1.4.0 compatible
 
 ## 2020-05-19